패키지 이름 설정, package-info.java 설명 추가
