--- src/controller/BookmarkController.java
@@ -1,20 +1,0 @@
-package controller;
-
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class BookmarkController extends HttpServlet {
-	
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException ServletException {
-		doPost(req, resp);
-	}
-
-	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException ServletException {
-		System.out.println("Servlet Loaded");
-		
-		req.setCharacterEncoding("utf-8");
-		resp.setCharacterEncoding("utf-8");
-	}
-
-}
