--- src/controller/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author student
- *
- */
-package controller;
