--- src/model/db/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author student
- *
- */
-package model.db;
