--- src/model/dto/Project.java
@@ -1,5 +1,0 @@
-package model.dto;
-
-public class Project {
-
-}
