--- src/model/dto/Tag.java
@@ -1,5 +1,0 @@
-package model.dto;
-
-public class Tag {
-
-}
