--- src/model/dto/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author student
- *
- */
-package model.dto;
