+++ src/portit/controller/FrontController.java
@@ -1,0 +1,21 @@
+package portit.controller;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet("/*")
+public class FrontController extends HttpServlet {
+	
+	@Override
+	protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		System.out.println("Servlet Loaded");
+		resp.setContentType("text/html; charset=UTF-8");		
+		
+	}
+
+}
