+++ src/portit/model/dao/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * DAO 클래스 패키지
+ * @auther gnsngck
+ */
+package portit.model.dao;
