+++ src/portit/model/db/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * DB 클래스 패키지
+ * @auther gnsngck
+ */
+package portit.model.db;
