+++ src/portit/model/dto/Bookmark.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Bookmark {
+
+}
