+++ src/portit/model/dto/Media.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Media {
+
+}
