+++ src/portit/model/dto/Member.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Member {
+
+}
