+++ src/portit/model/dto/Message.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Message {
+
+}
