+++ src/portit/model/dto/Notification.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Notification {
+
+}
