+++ src/portit/model/dto/Portfolio.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Portfolio {
+
+}
