+++ src/portit/model/dto/Profile.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Profile {
+
+}
