+++ src/portit/model/dto/Project.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Project {
+
+}
