+++ src/portit/model/dto/Tag.java
@@ -1,0 +1,5 @@
+package portit.model.dto;
+
+public class Tag {
+
+}
