+++ src/portit/model/dto/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * DTO/VO 클래스 패키지
+ * @auther gnsngck
+ */
+package portit.model.dto;
