+++ src/portit/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * PortIT
+ * @auther CaterJo, ksyeon, gnsngck, isyi, inaheaven, hyangkong
+ */
+package portit;
