+++ src/portit/view/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 뷰 클래스 패키지
+ * @auther gnsngck
+ */
+package portit.view;
